Fix CI failure: cast treatment to float before injecting non-integer D in test
`tests/test_chaisemartin_dhaultfoeuille.py::TestByPathNonBinary::test_non_integer_D_raises`
failed across all 7 CI runners (macos py3.11/3.14, ubuntu-24.04-arm
py3.11/3.13/3.14, ubuntu-latest py3.14, windows-latest py3.11) with:

    TypeError: Invalid value '1.5' for dtype 'int64'

The fixture builds `df["treatment"]` as integers in {0, 1, 2}, so the
column dtype is int64. On pandas >= 2.x (the version pinned in CI for
py3.11+ runners), `df.loc[mask, "treatment"] = 1.5` raises TypeError
outright; on older pandas (the local dev env) it emitted a
FutureWarning and proceeded. The test path was supposed to exercise
the estimator's D-integer guard at fit-time, not pandas's dtype
coercion behavior.

Cast `df["treatment"]` to float before injecting 1.5 so the test
behaves identically across pandas versions and reaches the
ValueError raise it was written to verify.

+++ tests/test_chaisemartin_dhaultfoeuille.py
@@ -8208,7 +8208,12 @@ class TestByPathNonBinary:
     def test_non_integer_D_raises(self):
         """D values containing 1.5 raise ValueError."""
         df = _by_path_data_with_non_binary_treatment()
-        # Inject a non-integer D for one cell
+        # Cast `treatment` to float BEFORE assigning 1.5: on pandas >= 2.x
+        # `df.loc[mask, "treatment"] = 1.5` raises `TypeError: Invalid
+        # value '1.5' for dtype 'int64'` outright instead of silently
+        # coercing. We want to inject a continuous value to test the
+        # estimator's D-integer guard, not exercise pandas dtype coercion.
+        df["treatment"] = df["treatment"].astype(float)
         mask = (df["group"] == 0) & (df["period"] == 4)
         df.loc[mask, "treatment"] = 1.5
         est = ChaisemartinDHaultfoeuille(
